// gospelee-admin/src/enums/account/AccountEcclesiaHistoryStatusType.ts
export enum AccountEcclesiaHistoryStatusType {
  JOIN_REQUEST = 'JOIN_REQUEST',      // 참여요청
  INVITE_REQUEST = 'INVITE_REQUEST',    // 초대요청
  JOIN_APPROVAL = 'JOIN_APPROVAL',     // 참여승인
  INVITE_APPROVAL = 'INVITE_APPROVAL',   // 초대승인
  JOIN_REJECT = 'JOIN_REJECT',     // 참여거절
  INVITE_REJECT = 'INVITE_REJECT',   // 초대거절
  LEAVE = 'LEAVE',
}